Put LaneCorridor definition in extra file
