+++ modules/world/map/road_corridor.hpp
@@ -14,8 +14,8 @@
 #include "modules/world/opendrive/opendrive.hpp"
 #include "modules/world/map/road.hpp"
 #include "modules/world/map/lane.hpp"
+#include "modules/world/map/lane_corridor.hpp"
 #include "modules/geometry/geometry.hpp"
-
 
 
 namespace modules {
@@ -25,56 +25,6 @@ namespace map {
 using modules::geometry::Line;
 using modules::geometry::Polygon;
 using modules::world::opendrive::XodrRoadId;
-
-
-struct LaneCorridor {
-  //! Getter
-  LanePtr& GetLeftLane(const LaneId& lane_id) {
-    // TODO(@hart): to be implemented
-  }
-  LanePtr& GetRightLane(const LaneId& lane_id) {
-    // TODO(@hart): to be implemented
-  }
-  Line& GetLeftBoundary() {
-    return left_boundary_;
-  }
-  Line& GetRightBoundary() {
-    return right_boundary_;
-  }
-  Line& GetCenterLine() {
-    return center_line_;
-  }
-  Polygon& GetMergedPolygon() {
-    return merged_polygon_;
-  }
-  std::map<float, LanePtr>& GetLanes() {
-    return lanes_;
-  }
-
-  //! Setter
-  void SetLeftBoundary(const Line& boundary) {
-    left_boundary_ = boundary;
-  }
-  void SetRightBoundary(const Line& boundary) {
-    right_boundary_ = boundary;
-  }
-  void SetCenterLine(const Line& line) {
-    center_line_ = line;
-  }
-  void SetMergedPolygon(const Polygon& poly) {
-    merged_polygon_ = poly;
-  }
-  void SetLane(float s_start, const LanePtr& lane) {
-    lanes_[s_start] = lane;
-  }
-
-  std::map<float, LanePtr> lanes_;  // s_end, LanePtr
-  Line center_line_;
-  Polygon merged_polygon_;
-  Line left_boundary_;
-  Line right_boundary_;
-};
-using LaneCorridorPtr = std::shared_ptr<LaneCorridor>;
 
 
 struct RoadCorridor {
